+++ app/src/org/commcare/views/widgets/AudioRecordingService.java
@@ -37,12 +37,11 @@ import java.io.IOException;
  * @author avazirna
  **/
 public class AudioRecordingService extends Service {
-    private static final int HEAAC_SAMPLE_RATE = 44100;
-    private static final int AMRNB_SAMPLE_RATE = 8000;
     private MediaRecorder recorder;
     private final IBinder binder = new AudioRecorderBinder();
     public static final String RECORDING_FILENAME_EXTRA_KEY = "recording-filename-extra-key";
     private NotificationManager notificationManager;
+    private AudioRecordingHelper audioRecordingHelper = new AudioRecordingHelper();
 
     @Override
     public void onCreate() {
@@ -59,18 +58,20 @@ public class AudioRecordingService extends Service {
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
         String fileName = intent.getExtras().getString(RECORDING_FILENAME_EXTRA_KEY);
-        setupRecorder(fileName);
+        if (recorder == null) {
+            recorder = audioRecordingHelper.setupRecorder(fileName);
+        }
         recorder.start();
         return START_NOT_STICKY;
     }
 
     @Override
     public void onDestroy() {
-        resetRecording();
+        resetRecorder();
         this.stopForeground(true);
     }
 
-    private void resetRecording() {
+    private void resetRecorder() {
         if (recorder != null) {
             recorder.release();
             recorder = null;
@@ -107,71 +108,6 @@ public class AudioRecordingService extends Service {
         public AudioRecordingService getService() {
             return AudioRecordingService.this;
         }
-    }
-
-    private void setupRecorder(String fileName) {
-        if (recorder == null) {
-            recorder = new MediaRecorder();
-        }
-
-        boolean isHeAacSupported = isHeAacEncoderSupported();
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
-            recorder.setAudioSource(MediaRecorder.AudioSource.VOICE_COMMUNICATION);
-        } else {
-            recorder.setAudioSource(MediaRecorder.AudioSource.MIC);
-        }
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
-            recorder.setPrivacySensitive(true);
-        }
-        recorder.setAudioSamplingRate(isHeAacSupported ? HEAAC_SAMPLE_RATE : AMRNB_SAMPLE_RATE);
-        recorder.setOutputFormat(MediaRecorder.OutputFormat.MPEG_4);
-        if (isHeAacSupported) {
-            recorder.setAudioEncoder(MediaRecorder.AudioEncoder.HE_AAC);
-        } else {
-            recorder.setAudioEncoder(MediaRecorder.AudioEncoder.AMR_NB);
-        }
-        recorder.setOutputFile(fileName);
-
-        try {
-            recorder.prepare();
-            Logger.log(LogTypes.TYPE_MEDIA_EVENT, "Preparing recording: " + fileName
-                    + " | " + (isHeAacSupported ? HEAAC_SAMPLE_RATE : AMRNB_SAMPLE_RATE)
-                    + " | " + (isHeAacSupported ? MediaRecorder.AudioEncoder.HE_AAC :
-                    MediaRecorder.AudioEncoder.AMR_NB));
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    // Checks whether the device supports High Efficiency AAC (HE-AAC) audio codec
-    private boolean isHeAacEncoderSupported() {
-        int numCodecs = MediaCodecList.getCodecCount();
-
-        for (int i = 0; i < numCodecs; i++) {
-            MediaCodecInfo codecInfo = MediaCodecList.getCodecInfoAt(i);
-
-            if (!codecInfo.isEncoder()) {
-                continue;
-            }
-
-            for (String supportedType : codecInfo.getSupportedTypes()) {
-                if (supportedType.equalsIgnoreCase(MIMETYPE_AUDIO_AAC)) {
-                    MediaCodecInfo.CodecCapabilities cap = codecInfo.getCapabilitiesForType(MIMETYPE_AUDIO_AAC);
-                    MediaCodecInfo.CodecProfileLevel[] profileLevels = cap.profileLevels;
-                    for (MediaCodecInfo.CodecProfileLevel profileLevel : profileLevels) {
-                        int profile = profileLevel.profile;
-                        if (profile == MediaCodecInfo.CodecProfileLevel.AACObjectHE
-                                || profile == MediaCodecInfo.CodecProfileLevel.AACObjectHE_PS) {
-                            return true;
-                        }
-                    }
-                }
-            }
-        }
-
-        return false;
     }
 
     @RequiresApi(api = Build.VERSION_CODES.Q)
